Rebuild the project using DSMR and MQTT Autodiscovery
The project is rebuilt with the DSMR library, adding
MQTT and autodiscovery on top.

By default all data from the P1 meter is sent to MQTT,
and all sensors are added automatically in HomeAssistant
using the MQTT auto discovery.

I wanted to read more data, especially the returned tariffs
and some other things, like voltage and current. But the
original code didn't allow this. Using the autodiscovery
you're getting a very easy integration, no need to edit
YAML files, and no need to restart HomeAssistant to are
or remove sensors.